test_push first two failure cases
