+++ ee/vellum_cli/tests/test_pull.py
@@ -1,19 +1,15 @@
-import pytest
 import io
 import os
-import shutil
 import tempfile
 from uuid import uuid4
 import zipfile
-from typing import Any, Callable, Dict, Generator, Tuple
 
 from click.testing import CliRunner
-import tomli_w
 
 from vellum_cli import main as cli_main
 
 
-def zip_file_map(file_map: dict[str, str]) -> bytes:
+def _zip_file_map(file_map: dict[str, str]) -> bytes:
     # Create an in-memory bytes buffer to store the zip
     zip_buffer = io.BytesIO()
 
@@ -29,44 +25,12 @@ def zip_file_map(file_map: dict[str, str]) -> bytes:
     return zip_bytes
 
 
-@pytest.fixture
-def mock_module() -> Generator[Tuple[str, str, Callable[[Dict[str, Any]], None]], None, None]:
-    current_dir = os.getcwd()
-    temp_dir = tempfile.mkdtemp()
-    os.chdir(temp_dir)
-    module = "examples.mock"
-
-    def set_pyproject_toml(vellum_config: Dict[str, Any]) -> None:
-        pyproject_toml_path = os.path.join(temp_dir, "pyproject.toml")
-        with open(pyproject_toml_path, "wb") as f:
-            tomli_w.dump(
-                {"tool": {"vellum": vellum_config}},
-                f,
-            )
-
-    set_pyproject_toml(
-        {
-            "workflows": [
-                {
-                    "module": module,
-                    "workflow_sandbox_id": str(uuid4()),
-                }
-            ]
-        }
-    )
-
-    yield temp_dir, module, set_pyproject_toml
-
-    os.chdir(current_dir)
-    shutil.rmtree(temp_dir)
-
-
 def test_pull(vellum_client, mock_module):
     # GIVEN a module on the user's filesystem
     temp_dir, module, _ = mock_module
 
     # AND the workflow pull API call returns a zip file
-    vellum_client.workflows.pull.return_value = iter([zip_file_map({"workflow.py": "print('hello')"})])
+    vellum_client.workflows.pull.return_value = iter([_zip_file_map({"workflow.py": "print('hello')"})])
 
     # WHEN the user runs the pull command
     runner = CliRunner()
@@ -87,7 +51,7 @@ def test_pull__second_module(vellum_client, mock_module):
     temp_dir, module, set_pyproject_toml = mock_module
 
     # AND the workflow pull API call returns a zip file
-    vellum_client.workflows.pull.return_value = iter([zip_file_map({"workflow.py": "print('hello')"})])
+    vellum_client.workflows.pull.return_value = iter([_zip_file_map({"workflow.py": "print('hello')"})])
 
     # AND the module we're about to pull is configured second
     set_pyproject_toml(
@@ -118,7 +82,7 @@ def test_pull__sandbox_id_with_no_config(vellum_client):
     workflow_sandbox_id = "87654321-0000-0000-0000-000000000000"
 
     # AND the workflow pull API call returns a zip file
-    vellum_client.workflows.pull.return_value = iter([zip_file_map({"workflow.py": "print('hello')"})])
+    vellum_client.workflows.pull.return_value = iter([_zip_file_map({"workflow.py": "print('hello')"})])
 
     # AND we are currently in a new directory
     current_dir = os.getcwd()
@@ -146,7 +110,7 @@ def test_pull__remove_missing_files(vellum_client, mock_module):
     temp_dir, module, _ = mock_module
 
     # AND the workflow pull API call returns a zip file
-    vellum_client.workflows.pull.return_value = iter([zip_file_map({"workflow.py": "print('hello')"})])
+    vellum_client.workflows.pull.return_value = iter([_zip_file_map({"workflow.py": "print('hello')"})])
 
     # AND there is already a different file in the module directory
     other_file_path = os.path.join(temp_dir, *module.split("."), "other_file.py")
@@ -175,7 +139,7 @@ def test_pull__remove_missing_files__ignore_pattern(vellum_client, mock_module):
     temp_dir, module, set_pyproject_toml = mock_module
 
     # AND the workflow pull API call returns a zip file
-    vellum_client.workflows.pull.return_value = iter([zip_file_map({"workflow.py": "print('hello')"})])
+    vellum_client.workflows.pull.return_value = iter([_zip_file_map({"workflow.py": "print('hello')"})])
 
     # AND there is already a different file in the module directory
     other_file_path = os.path.join(temp_dir, *module.split("."), "other_file.py")
@@ -227,7 +191,7 @@ def test_pull__include_json(vellum_client, mock_module):
 
     # AND the workflow pull API call returns a zip file
     vellum_client.workflows.pull.return_value = iter(
-        [zip_file_map({"workflow.py": "print('hello')", "workflow.json": "{}"})]
+        [_zip_file_map({"workflow.py": "print('hello')", "workflow.json": "{}"})]
     )
 
     # WHEN the user runs the pull command
@@ -249,7 +213,7 @@ def test_pull__exclude_code(vellum_client, mock_module):
 
     # AND the workflow pull API call returns a zip file
     vellum_client.workflows.pull.return_value = iter(
-        [zip_file_map({"workflow.py": "print('hello')", "workflow.json": "{}"})]
+        [_zip_file_map({"workflow.py": "print('hello')", "workflow.json": "{}"})]
     )
 
     # WHEN the user runs the pull command
