+++ tests/workflows/basic_emitter_workflow/tests/test_workflow.py
@@ -1,10 +1,8 @@
 import pytest
 from datetime import datetime
-import json
 import time
 
 from vellum.workflows.events.types import default_serializer
-from vellum.workflows.state.encoder import DefaultStateEncoder
 
 from tests.workflows.basic_emitter_workflow.workflow import BasicEmitterWorkflow, ExampleEmitter, NextNode, StartNode
